refactor: 파일 분리 및 코드 일부 리팩터링
